requirements.txt
Flask==2.3.2
numpy==1.23.3
Pillow==9.5.0
tensorboardX==2.6
timm==0.6.13
torch==1.11.0+cu115
torchvision==0.12.0
tqdm==4.63.0


# server public IP: 3.6.164.8
# user name: ubuntu